UWP tests can now activate context menus
