+++ Xamarin.Forms.Core.Windows.UITests/Xamarin.Forms.Core.Windows.UITests.csproj
@@ -74,7 +74,9 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WindowsTestBase.cs" />
+    <Compile Include="WindowsTestServer.cs" />
     <Compile Include="WinDriverApp.cs" />
+    <Compile Include="WinQuery.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
